Add absolute targets

Set targets for a project prioritization problem by specifying exactly what probability of persistence is required for each feature. For instance, setting an absolute target of 10% (i.e. 0.1) corresponds to a threshold 10% probability of persisting.

# S4 method for ProjectProblem,numeric
add_absolute_targets(x, targets)

# S4 method for ProjectProblem,character
add_absolute_targets(x, targets)

Arguments

x
ProjectProblem-class object.

targets
Object that specifies the targets for each feature. See the Details section for more information.

Details

Targets are used to specify the minimum probability of persistence for each feature in solutions. For minimum set objectives (i.e. add_min_set_objective, these targets specify the minimum probability of persistence required for each species in the solution. And for budget constrained objectives that use targets (i.e.add_max_targets_met_objective), these targets specify the minimum threshold probability of persistence that needs to be achieved to count the benefits for conserving these species. Please note that attempting to solve problems with objectives that require targets without specifying targets will throw an error.

The targets for a problem can be specified in several different ways:

numeric

vector of target values for each feature. The order of the target values should correspond to the order of the features in the data used to create the argument to x. Additionally, for convenience, this type of argument can be a single value to assign the same target to each feature.

character

specifying the name of column in the feature data (i.e. the argument to features in the problem function) that contains the persistence targets.

See also

Examples

#TODO