Add feature weights

Set weights for conserving features in a project prioritization problem.

# S4 method for ProjectProblem,numeric
add_feature_weights(x, weights)

# S4 method for ProjectProblem,character
add_feature_weights(x, weights)

Arguments

x
ProjectProblem-class object.

weights
Object that specifies the weights for each feature. See the Details section for more information.

Details

Weights are used to specify the relative importance for maintaining the persistence of specific features. For budget constrained problems (e.g. add_max_sum_persistence_objective), these weights could be used to specify which features are more important than other features according to evolutionary or cultural metrics. Specifically, features with a higher weight value are considered more important. It is generally best to ensure that the feature weights range between 0.0001 and 10,000 to reduce the time required to solve problems using exact algorithm solvers. As a consequence, you might have to rescale the feature weights if the largest or smallest values occur outside this range (excluding zeros). If you want to ensure that certain features conserved in the solutions, it is strongly recommended to lock in the actions for these features instead of setting really high weights for these features. Please note that a warning will be thrown if you attempt to solve problems with weights when an objective has been specified that does not use weights. Currently, all objectives---except for the minimum set objective (i.e. add_min_set_objective)---can use weights.

The weights for a problem can be specified in several different ways:

numeric

vector of weight values for each feature.

character

specifying the name of column in the feature data (i.e. the argument to features in the problem function) that contains the weights.

See also

Examples

#TODO