Problem objective

An objective is used to specify the overall goal of a project prioritization problem. All project prioritization problems involve minimizing or maximizing some kind of objective. For instance, the decision maker may require a funding scheme that maximizes the total number of species that are expected to persist into the future whilst ensuring that the total cost of the funded actions does not exceed a budget. Alternatively, the planner may require a solution that ensures that each species meets a target level of persistence whilst minimizing the cost of the funded actions.

Details

Please note that failing to specify an objective before attempting to solve a problem will return an error.

The following objectives can be added to a conservation planning problem:

add_min_set_objective

Minimize the cost of the solution whilst ensuring that all targets are met. This objective is conceptually similar to that used in Marxan---except that decisions are made on funding actions instead of purchasing areas for conservation.

add_max_persistence_objective

Maximize the probability that at last one feature will persist into the future.

add_max_richness_objective

Maximize the expected richness of the features. In other words, maximize the total number of species that are expected to persist.

add_max_targets_met_objective

Maximize the number of persistence targets that are met for the features.

add_max_phylo_div_objective

Maximize the expected phylogenetic diversity of the features.

See also

Examples

#TODO